Automatically rebuild if *any* file changes
Previous to this commit, the auto rebuilder would only rebuild
individual files that have been changed. However, HAML/SCSS/SASS files
can have import dependencies on eachother. Modifying a file which is
a dependency of another file did not cause dependent file to rebuild,
meaning that the emitted resources did not stay up to date.

Ideally, the builder would be capable of tracking import dependencies
and only rebuilding the minimal set of files necessary to satisfy
updates anywhere in the dependency chain. However, this is impossible to
do correctly without detailed integration with the underlying rendering
engine.

In the absence of this ideal solution, this commit modifies to
auto-builder to rebuild *all* files, if *any* files have changed. Due to
the speed of compilation in the common case, this shouldn't cause any
problems, and ensures thatthe auto-builder is still useful even when
modifying dependencies.